third_party/blink/web_tests/fast/animation/scroll-animations/scrolltimeline-creation-no-document.html

<!DOCTYPE html>
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>
<script>
test(function() {
  document.documentElement.remove();
  const timeline = new ScrollTimeline({timeRange: 100});
  assert_equals(timeline.scrollSource, null);
  assert_equals(timeline.currentTime, null);
}, 'The scrollSource can be null, if the document.scrollingElement does not exist');
</script>
